refactor(github): Refactored the `GithubService` to be a factory instead of a singleton.
With this patch the `GithubService` is no longer a singleton and
does not need any configuration within the configuration phase of
your app.

Instead it is now a factory that returns a new module. The module
exposes the same methods the singleton did (`getIssuesForMilestone`,
`getMilestone`, `getMilestones`). The `owner`, `repo` and `token` values
are not exposed.