feat(toggle): Introduce `[ed-toggle-attr]`.
Add attribute directive `[ed-toggle-attr]` to toggle attributes. The
directive only toggles attributes, NOT their values!

It is possible to create toggle groups with `ed-toggle-attr-group="<name>"`.
Within a group there will only be one element with the specified
attribute. This option can be used for dropdowns and collapsable
elements.

With the `ed-toggle-attr-closest="<tag>"` attribute it is possible to
toggle attributes not for the current element but one of its parents.
Note that parent selection is only support with tag names no CSS selectors.