feat(github): Fetch additional information for PRs.
Github doesn't send much information about issues if they are also
an PR. But it does provide the API to get more information inside
the `pull_request` property.

In order to display if a PR is already merged we need this data.
So we pull it with an additional request.