feat(pull-request): Introduce module for pull requests.
The module exposes a `<ed-pull-request-list>` element to capsulate
a list of (!!!) pull requests! The attached controller also publishs
a metho to retrieve the state of a PR.

This patch also contains the the `<ed-badge>` element. This custom
component does NOT have any logic to it. It's just easier to style a
custom element instead of more divs...
The badge is used to display the current state of a PR to the user.
Red -> closesd, green -> open, purple -> merged.