feat(dashboard): Group milestones by tag.
Allow ed to group milestones by tags (title starting with `[<tag>]`).
The groups can be configured via the `grunt config` command. If the
milestone has noe tag it will be assigned to a "fallback" group.
If no groups are specified, mielstones will be put into a "milestones"
group.

Note that we could generate this grouping automatically by iterating over
all milestones titles. But this would not allow the user to configure
an ordering of the groups.