feat(github): First version of `GithubStore`.
The `GithubStore` service will be use as a central source for ED to get
its data. Everything that is displayed should the piped through this
service. This allows us to have more or less a unidirectional data flow.

This service will be responsible to keep our data fresh. Meaning, you can
add repositories. When you fetch milestones from that repository the service
will check if it has fresh enough data. If not it will fetch again from
Github.